submits/400422202/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.13
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
ipython==7.16.3
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==3.0.3
Khayyam==3.0.17
MarkupSafe==2.0.1
numpy==1.19.5
openapi-codec==1.3.2
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.5.4
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
